% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_of_mome.R
\name{coef.index_of_mome}
\alias{coef.index_of_mome}
\title{Print the Output of 'index_of_mome'}
\usage{
\method{coef}{index_of_mome}(object, ...)
}
\arguments{
\item{object}{The output of \code{\link[=index_of_mome]{index_of_mome()}}.}

\item{...}{Optional arguments. Ignored.}
}
\value{
Scalar: The index of moderated-mediation in \code{object}.
}
\description{
Extract the Index of Moderated-Mediation.
}
\details{
The \code{coef} method of the \code{index_of_mome}-class object.
}
\seealso{
\code{\link[=index_of_mome]{index_of_mome()}}
}
