% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_effect.R
\name{cond_effect}
\alias{cond_effect}
\title{One Line Title}
\usage{
cond_effect(fit, x, y, m = NULL, wvalues = NULL, xvalues = c(-1, 1))
}
\arguments{
\item{fit}{The fit object. Currently only supports a
\link[lavaan:lavaan-class]{lavaan::lavaan} object.}

\item{x}{Character. The name of predictor at the start of the pathway.}

\item{y}{Character. The name of the outcome variable at
the end of the pathway.}

\item{m}{A vector of the variable names of the
moderators. The pathway goes from the first
mediator successively to the last mediator. If
\code{NULL}, the default, the pathway goes from \code{x}
to \code{y}.}

\item{wvalues}{A numeric vector of named elements. The names are the variable
names of the moderators, and the values are the values to
which the moderators will be set to. Default is \code{NULL}.}

\item{xvalues}{A vector of two numbers. The values of \code{y}
predicted by these two values of \code{x} will
be computed. Default is \code{c(-1, 1)}.}
}
\value{
Specify what are returned.
}
\description{
One paragraph description
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{
\donttest{
}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
