% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_of_mome.R
\name{confint.index_of_mome}
\alias{confint.index_of_mome}
\title{Confidence Interval of the Output of 'index_of_mome'}
\usage{
\method{confint}{index_of_mome}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{The output of \code{\link[=index_of_mome]{index_of_mome()}}.}

\item{parm}{Ignored.}

\item{level}{The level of confidence for the bootstrap confidence
interval. Default is .95. Must match the level of the stored
confidence interval.}

\item{...}{Optional arguments. Ignored.}
}
\value{
A one-row-two-column data frame of the confidence limits.
}
\description{
Extract the confidence interval the output of \code{\link[=index_of_mome]{index_of_mome()}}.
}
\details{
The \code{confint} method of the \code{index_of_mome}-class object.
}
