% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_indirect.R
\name{confint.indirect}
\alias{confint.indirect}
\title{Confidence Interval of
Indirect Effect or Conditional
Indirect Effect}
\usage{
\method{confint}{indirect}(object, parm, level = NULL, boot_type, ...)
}
\arguments{
\item{object}{The output of
\code{\link[=indirect_effect]{indirect_effect()}} or
\code{\link[=cond_indirect]{cond_indirect()}}.}

\item{parm}{Ignored because the
stored object always has only one
parameter.}

\item{level}{If set to \code{NULL}, the default,
then the
level of confidence used to generate
\code{object}
is used. If set to a value,
this value will be used
to recompute the confidence intervals.
If the confidence interval is to be
computed from the standard error,
and so \code{level} is not set in \code{object},
then the default value is .95.
(This new behavior applies to 0.3.6.15
and later version.)}

\item{boot_type}{If bootstrap
confidence interval is to be formed,
the type of bootstrap confidence
interval. The supported types
are \code{"perc"} (percentile bootstrap
confidence interval, the recommended
method) and \code{"bc"}
(bias-corrected, or BC, bootstrap
confidence interval). If not supplied,
the stored \code{boot_type} will be used.}

\item{...}{Additional arguments.
Ignored by the function.}
}
\value{
A numeric vector of
two elements, the limits of
the confidence interval.
}
\description{
Return the
confidence interval of the indirect
effect or conditional indirect effect
stored in the output of
\code{\link[=indirect_effect]{indirect_effect()}} or
\code{\link[=cond_indirect]{cond_indirect()}}.
}
\details{
It extracts and returns the
stored confidence interval
if available.

The type of confidence interval
depends on the call used to
compute the effect. This function
merely retrieves the stored estimates,
which could be generated by
nonparametric bootstrapping,
Monte Carlo simulation, or other
methods to be supported in
the future, and uses them to form the
percentile confidence interval.

If the following conditions are met, the
stored standard errors, if available,
will be used test an effect and
form it confidence interval:
\itemize{
\item Confidence intervals have not been
formed (e.g., by bootstrapping or
Monte Carlo).
\item The path has no mediators.
\item The model has only one group.
\item The path is moderated by one or
more moderator.
\item Both the \code{x}-variable and the
\code{y}-variable are not standardized.
}

If the model is fitted by OLS
regression (e.g., using \code{\link[stats:lm]{stats::lm()}}),
then the variance-covariance matrix
of the coefficient estimates will be
used, and confidence
intervals are computed from the \emph{t}
statistic.

If the model is fitted by structural
equation modeling using \code{lavaan}, then
the variance-covariance computed by
\code{lavaan} will be used,
and confidence intervals are computed
from the \emph{z} statistic.
\subsection{Caution}{

If the model is fitted by structural
equation modeling and has moderators,
the standard errors, \emph{p}-values,
and confidence interval computed
from the variance-covariance matrices
for conditional effects
can only be trusted if all covariances
involving the product terms are free.
If any of them are fixed, for example,
fixed to zero, it is possible
that the model is not invariant to
linear transformation of the variables.
}
}
\examples{

dat <- modmed_x1m3w4y1

# Indirect Effect

library(lavaan)
mod1 <-
"
m1 ~ x
m2 ~ m1
y  ~ m2 + x
"
fit <- sem(mod1, dat,
           meanstructure = TRUE, fixed.x = FALSE,
           se = "none", baseline = FALSE)
# R should be at least 2000 or 5000 in real research.
out1 <- indirect_effect(x = "x", y = "y",
                        m = c("m1", "m2"),
                        fit = fit,
                        boot_ci = TRUE, R = 40, seed = 2345,
                        parallel = FALSE,
                        progress = FALSE)
out1
confint(out1)


}
\seealso{
\code{\link[=indirect_effect]{indirect_effect()}} and
\code{\link[=cond_indirect]{cond_indirect()}}
}
