% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_7_mod.R
\docType{data}
\name{data_mod_2x2w}
\alias{data_mod_2x2w}
\title{Sample Dataset: Two Moderators on Two Predictors}
\format{
A data frame with 200 rows
and 6 variables:
\describe{
\item{y}{Outcome variable. Numeric.}
\item{x1}{Predictor 1. Numeric.}
\item{x2}{Predictor 2. Numeric.}
\item{w1}{Moderator 1. Numeric.}
\item{w2}{Moderator 2. Numeric.}
\item{c1}{Control variable. Numeric.}
\item{c2}{Control variable. Numeric.}
}
}
\usage{
data_mod_2x2w
}
\description{
A two-moderator model,
with the moderators affecting the
effects of different predictors.
}
\examples{
library(lavaan)
data(data_mod_2x2w)
lm_out <- lm(y ~ x1*w1 + x2*w2 + c1 + c2, data_mod_2x2w)
out1 <- cond_effects(
  wlevels = "w1",
  x = "x1",
  fit = lm_out
)
out1
plot(out1, graph_type = "tumble")

out2 <- cond_effects(
  wlevels = "w2",
  x = "x2",
  fit = lm_out
)
out2
plot(out2, graph_type = "tumble")
}
\keyword{datasets}
