% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_4.R
\docType{data}
\name{data_mod_cat}
\alias{data_mod_cat}
\title{Moderation with One Categorical Moderator: Sample Dataset}
\format{
A data frame with 300 rows and 5 variables:
\describe{
\item{x}{Predictor. Numeric.}
\item{w}{Moderator. String. Values: "group1", "group2", "group3"}
\item{y}{Outcome variable. Numeric.}
\item{c1}{Control variable. Numeric.}
\item{c2}{Control variable. Numeric.}
}
}
\usage{
data_mod_cat
}
\description{
A moderation model with a categorical moderator.
}
\examples{
data(data_mod_cat)
dat <- data_mod_cat
summary(lm_y <- lm(y ~ x*w + c1 + c2, dat))
}
\keyword{datasets}
