% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect.R
\name{indirect}
\alias{indirect}
\title{One Line Title}
\usage{
indirect(x, y, m = NULL, fit = NULL, est = NULL, wvalues = NULL)
}
\arguments{
\item{x}{Character. The name of predictor at the start of the pathway.}

\item{y}{Character. The name of the outcome variable at
the end of the pathway.}

\item{m}{A vector of the variable names of the
moderators. The pathway goes from the first
mediator successively to the last mediator. If
\code{NULL}, the default, the pathway goes from \code{x}
to \code{y}.}

\item{fit}{The fit object. Currently only supports a
\link[lavaan:lavaan-class]{lavaan::lavaan} object.}

\item{est}{The output of \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. If \code{NULL}, the
default, it will be generated from \code{fit}. If supplied,
\code{fit} will ge ignored.}

\item{wvalues}{A numeric vector of named elements. The names are the variable
names of the moderators, and the values are the values to
which the moderators will be set to. Default is \code{NULL}.}
}
\value{
Specify what are returned.
}
\description{
One paragraph description
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{

library(lavaan)
dat <- modmed_x1m3w4y1
mod <-
"
m1 ~ a1 * x   + b1 * w1 + d1 * x:w1
m2 ~ a2 * m1  + b2 * w2 + d2 * m1:w2
m3 ~ a3 * m2  + b3 * w3 + d3 * m2:w3
y  ~ a4 * m3  + b4 * w4 + d4 * m3:w4
"
fit <- sem(mod, dat, meanstructure = TRUE, fixed.x = FALSE, se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

wvalues <- c(w1 = 5, w2 = 4, w3 = 2, w4 = 3)

indirect_1 <- indirect(x = "x", y = "y", m = c("m1", "m2", "m3"), fit = fit,
                       wvalues = wvalues)
indirect_2 <- (est[est$label == "a1", "est"] +
                wvalues["w1"] * est[est$label == "d1", "est"]) *
              (est[est$label == "a2", "est"] +
                wvalues["w2"] * est[est$label == "d2", "est"]) *
              (est[est$label == "a3", "est"] +
                wvalues["w3"] * est[est$label == "d3", "est"]) *
              (est[est$label == "a4", "est"] +
                wvalues["w4"] * est[est$label == "d4", "est"])
indirect_1
indirect_2

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
