% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_indirect_list.R
\name{indirect_effects_from_list}
\alias{indirect_effects_from_list}
\title{Coefficient Table of an 'indirect_list' Class
Object}
\usage{
indirect_effects_from_list(object, add_sig = TRUE)
}
\arguments{
\item{object}{The output of
\code{\link[=indirect_effect]{indirect_effect()}} or
\code{\link[=cond_indirect]{cond_indirect()}}.}

\item{add_sig}{Whether a column
of significance test results
will be added. Default is \code{TRUE}.}
}
\value{
A data frame with the
indirect effect estimates and
confidence interval (if available).
}
\description{
Create a coefficient table
for the point estimates and
confidence intervals (if available)
for the
output of \code{\link[=many_indirect_effects]{many_indirect_effects()}}.
}
\examples{

# Create a data frame of the indirect effect estimates

out_df <- indirect_effects_from_list(out)
out_df


}
\seealso{
\code{\link[=many_indirect_effects]{many_indirect_effects()}}
}
