% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_data_used.R
\name{lav_data_used}
\alias{lav_data_used}
\title{One Line Title}
\usage{
lav_data_used(fit, drop_colon = TRUE)
}
\arguments{
\item{fit}{The fit object. Currently only supports a
\link[lavaan:lavaan-class]{lavaan::lavaan} object.}

\item{drop_colon}{Logical. Whether product terms (of the form "x:y")
will be dropped. Default is \code{TRUE}.}
}
\value{
Specify what are returned.
}
\description{
One paragraph description
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{
\donttest{
}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
