% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2ptable.R
\name{lm2ptable}
\alias{lm2ptable}
\title{Convert a 'lm' Output to a 'lavaan'-Like Object}
\usage{
lm2ptable(outputs)
}
\arguments{
\item{outputs}{A list of \code{lm} class objects.}
}
\value{
A list with the following arguments:
\itemize{
\item \code{est}: A data frame similar to the parameter estimates table
of the output of \code{\link[lavaan:sem]{lavaan::sem()}}.
\item \code{dat}: A data frame with all the data in the regression
models merged.
\item \code{implied_stats}: A list with two elements: \code{cov} is the
implied covariance matrix and \code{mean} is the implied
means. They are used to compute standardized
effects or form levels of moderators.
}
}
\description{
It only convert information necessary for
\code{\link[=cond_indirect]{cond_indirect()}} to compute conditional effects or
conditional indirect effects.
}
\details{
Users usually do not need to call this function
to use \code{\link[=cond_indirect]{cond_indirect()}} and \code{\link[=cond_indirect_effects]{cond_indirect_effects()}}.
These two functions will do the conversion internally
if necessary.
}
\examples{

data(data_serial)
dat <- data_serial
lm_m1 <- lm(m1 ~ x + c1 + c2, dat)
lm_m2 <- lm(m2 ~ m1 + x + c1 + c2, dat)
lm_y <- lm(y ~ m2 + m1 + x + c1 + c2, dat)
out <- lm2ptable(list(lm_m1, lm_m2, lm_y))
out$est


}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
