% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan2lm_list.R
\name{lm_from_lavaan_list}
\alias{lm_from_lavaan_list}
\title{Convert a 'lavaan' Model to an 'lm_from_lavaan_list'-Class Object}
\usage{
lm_from_lavaan_list(fit)
}
\arguments{
\item{fit}{A \code{lavaan}-class object, usually the output of \code{\link[lavaan:lavaan]{lavaan::lavaan()}}
or its wrappers.}
}
\value{
A \code{lm_from_lavaan_list}-class object, which is a list of \code{lm_from_lavaan}
objects. It has a \code{predict}-method (\code{\link[=predict.lm_from_lavaan_list]{predict.lm_from_lavaan_list()}}) for
computing the predicted values from one variable to another.
}
\description{
Converts the regression models in a \verb{lavaan`` model to an }lm_from_lavaan_list`-class object.
}
\details{
It identifies all dependent variables in a \code{lavaan} model
and creates a \code{lm_from_lavaan}-class object for each of them.
}
\examples{
library(lavaan)
data(data_med)
mod <-
"
m ~ a * x + c1 + c2
y ~ b * m + x + c1 + c2
"
fit <- sem(mod, data_med, fixed.x = FALSE)
fit_list <- lm_from_lavaan_list(fit)
tmp <- data.frame(x = 1, c1 = 2, c2 = 3, m = 4)
predict(fit_list, x = "x", y = "y", m = "m", newdata = tmp)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
