% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manymome-package.R
\docType{package}
\name{manymome-package}
\alias{manymome}
\alias{manymome-package}
\title{manymome: Mediation, Moderation and Moderated-Mediation After Model Fitting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Computes indirect effects, conditional effects, and conditional indirect effects in a structural equation model or path model after model fitting, with no need to define any user parameters or label any paths in the model syntax, using the approach presented in Cheung and Cheung (2023) \doi{10.3758/s13428-023-02224-z}. Can also form bootstrap confidence intervals by doing bootstrapping only once and reusing the bootstrap estimates in all subsequent computations. Supports bootstrap confidence intervals for standardized (partially or completely) indirect effects, conditional effects, and conditional indirect effects as described in Cheung (2009) \doi{10.3758/BRM.41.2.425} and Cheung, Cheung, Lau, Hui, and Vong (2022) \doi{10.1037/hea0001188}. Model fitting can be done by structural equation modeling using lavaan() or regression using lm().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/manymome/}
  \item Report bugs at \url{https://github.com/sfcheung/manymome/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Sing-Hang Cheung (\href{https://orcid.org/0000-0001-5182-0752}{ORCID})
}

}
\keyword{internal}
