% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmed_x1m3w4y1.R
\docType{data}
\name{modmed_x1m3w4y1}
\alias{modmed_x1m3w4y1}
\title{Sample Dataset: Moderated Serial Mediation}
\format{
A data frame with 200 rows and 11 variables:
\describe{
\item{x}{Predictor. Numeric.}
\item{w1}{Moderator 1. Numeric.}
\item{w2}{Moderator 2. Numeric.}
\item{w3}{Moderator 3. Numeric.}
\item{w4}{Moderator 4. Numeric.}
\item{m1}{Mediator 1. Numeric.}
\item{m2}{Mediator 2. Numeric.}
\item{m3}{Mediator 3. Numeric.}
\item{y}{Outcome variable. Numeric.}
\item{gp}{Three values: "blue", "yellow", "green". String.}
\item{city}{Four values: "alpha", "beta", "gamma", "sigma". String.}
}
}
\usage{
modmed_x1m3w4y1
}
\description{
Generated from a serial mediation model with
one predictor, three mediators, and one outcome
variable, with one moderator in each stage.
}
\keyword{datasets}
