% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_2_and_3.R
\docType{data}
\name{one_moderator}
\alias{one_moderator}
\title{One Moderator: Sample Dataset}
\format{
A data frame with 100 rows and 5 variables:
\describe{
\item{x}{Predictor. Numeric.}
\item{w}{Moderator. Numeric.}
\item{y}{Outcome variable. Numeric.}
\item{c1}{Control variable. variable. Numeric.}
\item{c2}{Control variable. Numeric.}
}
}
\usage{
one_moderator
}
\description{
A one-moderator model.
}
\examples{
library(lavaan)
data(one_moderator)
one_moderator$xw <- one_moderator$x * one_moderator$w
mod <-
"
y ~ a * x + w + d * xw + c1 + c2
w ~~ v_w * w
w ~ m_w * 1
a_lo := a + d * (m_w - sqrt(v_w))
a_hi := a + d * (m_w + sqrt(v_w))
"
fit <- sem(mod, one_moderator, fixed.x = FALSE)
parameterEstimates(fit)[c(1, 3, 6, 7, 24, 25), ]
}
\keyword{datasets}
