% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_i.R
\name{predict_i}
\alias{predict_i}
\title{One Line Title}
\usage{
predict_i(x, y, m = NULL, xvalue, fit)
}
\arguments{
\item{x}{Character. The name of predictor at the start of the pathway.}

\item{y}{Character. The name of the outcome variable at
the end of the pathway.}

\item{m}{A vector of the variable names of the
moderators. The pathway goes from the first
mediator successively to the last mediator. If
\code{NULL}, the default, the pathway goes from \code{x}
to \code{y}.}

\item{xvalue}{Scalar. The value of \code{y} predicted by this
value of \code{x} will be computed.}

\item{fit}{The fit object. Currently only supports a
\link[lavaan:lavaan-class]{lavaan::lavaan} object.}
}
\value{
Specify what are returned.
}
\description{
One paragraph description
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{
\donttest{
}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
