% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_of_mome.R
\name{print.index_of_mome}
\alias{print.index_of_mome}
\title{Print the Output of 'index_of_mome'}
\usage{
\method{print}{index_of_mome}(x, digits = 3, ...)
}
\arguments{
\item{x}{The output of \code{\link[=index_of_mome]{index_of_mome()}}.}

\item{digits}{The number of decimal places in the printout.}

\item{...}{Optional arguments. Ignored.}
}
\value{
Return \code{x} invisibly. Called for its side effect.
}
\description{
Print the output of \code{\link[=index_of_mome]{index_of_mome()}}.
}
\details{
The \code{print} method of the \code{index_of_mome}-class object.
}
\seealso{
\code{\link[=index_of_mome]{index_of_mome()}}
}
