% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_1_i.R
\name{refit_1_i}
\alias{refit_1_i}
\title{One Line Title}
\usage{
refit_1_i(x, xvalue, fit, wvalues = NULL)
}
\arguments{
\item{x}{Character. The name of predictor at the start of the pathway.}

\item{xvalue}{Scalar. The value of \code{y} predicted by this
value of \code{x} will be computed.}

\item{fit}{The fit object. Currently only supports a
\link[lavaan:lavaan-class]{lavaan::lavaan} object.}

\item{wvalues}{A numeric vector of named elements. The names are the variable
names of the moderators, and the values are the values to
which the moderators will be set to. Default is \code{NULL}.}
}
\value{
Specify what are returned.
}
\description{
One paragraph description
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{
\donttest{
}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
