% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_4.R
\docType{data}
\name{serial_parallel_mediation_categorical_moderators}
\alias{serial_parallel_mediation_categorical_moderators}
\title{Serial-Parallel Moderated Mediation with Two Categorical Moderators:
Sample Dataset}
\format{
A data frame with 300 rows and 8 variables:
\describe{
\item{x}{Predictor. Numeric.}
\item{w1}{Moderator. String. Values: "group1", "group2", "group3"}
\item{w2}{Moderator. String. Values: "team1", "team2"}
\item{m1}{Mediator 1. Numeric.}
\item{m2}{Mediator 2. Numeric.}
\item{y}{Outcome variable. Numeric.}
\item{c1}{Control variable. Numeric.}
\item{c2}{Control variable. Numeric.}
}
}
\usage{
serial_parallel_mediation_categorical_moderators
}
\description{
A serial-parallel mediation model with two categorical
moderators.
}
\examples{
data(serial_parallel_mediation_categorical_moderators)
dat <- serial_parallel_mediation_categorical_moderators
summary(lm_m11 <- lm(m11 ~ x*w1 + c1 + c2, dat))
summary(lm_m12 <- lm(m12 ~ m11 + x + w1 + c1 + c2, dat))
summary(lm_m2 <- lm(m2 ~ x + w1 + c1 + c2, dat))
summary(lm_y <- lm(y ~ m12 + m2*w2 + m12 + x + c1 + c2, dat))
}
\keyword{datasets}
