% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_2_and_3.R
\docType{data}
\name{simple_mediation_ab_moderated_one_moderator}
\alias{simple_mediation_ab_moderated_one_moderator}
\title{Simple Mediation with Both Paths Moderated By a Moderator: Sample Dataset}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{x}{Predictor. Numeric.}
\item{w}{Moderator. Numeric.}
\item{m}{Mediator. Numeric.}
\item{y}{Outcome variable. Numeric.}
\item{c1}{Control variable. variable. Numeric.}
\item{c2}{Control variable. Numeric.}
}

A data frame with 100 rows and 6 variables:
\describe{
\item{x}{Predictor. Numeric.}
\item{w}{Moderator. Numeric.}
\item{m}{Mediator. Numeric.}
\item{y}{Outcome variable. Numeric.}
\item{c1}{Control variable. variable. Numeric.}
\item{c2}{Control variable. Numeric.}
}
}
\usage{
simple_mediation_ab_moderated_one_moderator

simple_mediation_ab_moderated_one_moderator
}
\description{
A simple mediation model with a-path and b-path
moderated by one moderator.

A simple mediation model with a-path and b-path
moderated by one moderator.
}
\examples{
library(lavaan)
data(simple_mediation_ab_moderated_one_moderator)
simple_mediation_ab_moderated_one_moderator$xw <-
 simple_mediation_ab_moderated_one_moderator$x *
 simple_mediation_ab_moderated_one_moderator$w
simple_mediation_ab_moderated_one_moderator$mw <-
 simple_mediation_ab_moderated_one_moderator$m *
 simple_mediation_ab_moderated_one_moderator$w
mod <-
"
m ~ a * x + w + da * xw + c1 + c2
y ~ b * m + x + w + db * mw + c1 + c2
w ~~ v_w * w
w ~ m_w * 1
ab := a * b
ab_lo := (a + da * (m_w - sqrt(v_w))) * (b + db * (m_w - sqrt(v_w)))
ab_hi := (a + da * (m_w + sqrt(v_w))) * (b + db * (m_w + sqrt(v_w)))
"
fit <- sem(mod, simple_mediation_ab_moderated_one_moderator,
           meanstructure = TRUE, fixed.x = FALSE)
parameterEstimates(fit)[c(1, 3, 6, 9, 38:40), ]
library(lavaan)
data(simple_mediation_ab_moderated_one_moderator)
simple_mediation_ab_moderated_one_moderator$xw <-
 simple_mediation_ab_moderated_one_moderator$x *
 simple_mediation_ab_moderated_one_moderator$w
simple_mediation_ab_moderated_one_moderator$mw <-
 simple_mediation_ab_moderated_one_moderator$m *
 simple_mediation_ab_moderated_one_moderator$w
mod <-
"
m ~ a * x + w + da * xw + c1 + c2
y ~ b * m + x + w + db * mw + c1 + c2
w ~~ v_w * w
w ~ m_w * 1
ab := a * b
ab_lo := (a + da * (m_w - sqrt(v_w))) * (b + db * (m_w - sqrt(v_w)))
ab_hi := (a + da * (m_w + sqrt(v_w))) * (b + db * (m_w + sqrt(v_w)))
"
fit <- sem(mod, simple_mediation_ab_moderated_one_moderator,
           meanstructure = TRUE, fixed.x = FALSE)
parameterEstimates(fit)[c(1, 3, 6, 9, 38:40), ]
}
\keyword{datasets}
