% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmodsem-package.R
\docType{package}
\name{stdmodsem-package}
\alias{stdmodsem}
\alias{stdmodsem-package}
\title{stdmodsem: Moderation, Mediation, and Moderated-Mediation in SEM}
\description{
Functions for computing indirect effects, moderation effects, and conditional indirect effects, and their bootstraping confidence interval. Support standadized (partially or completely) moderation effects and conditional indirect effects. For standardized effects that involve the moderator, the sampling variation in the standardization will be taken into account.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/stdmodsem/}
  \item Report bugs at \url{https://github.com/sfcheung/stdmodsem/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

}
\keyword{internal}
