% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmodsem-package.R
\docType{package}
\name{stdmodsem}
\alias{stdmodsem}
\title{stdmodsem: Functions for computing indirect effects,
moderation effects, and conditional indirect effects, and
their bootstraping confidence interval. Support
standadized (partially or completely) moderation effects
and conditional indirect effects. For standardized
effects that involve the moderator, the sampling
variation in the standardization will be taken into
account.}
