% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan2lm_list.R
\name{terms.lm_from_lavaan}
\alias{terms.lm_from_lavaan}
\title{Model Terms of a 'lm_from_lavaan'-Class Object}
\usage{
\method{terms}{lm_from_lavaan}(x, ...)
}
\arguments{
\item{x}{A 'lm_from_lavaan'-class object.}

\item{...}{Additional arguments. Ignored.}
}
\value{
A \code{terms}-class object. See \link{terms.object} for details.
}
\description{
Extract the terms object rom a \code{lm_from_lavaan}-class
object.
}
\details{
A \code{lm_from_lavaan}-class object converts a regression model
for a variable in a \verb{lavaan`` model to a }formula\verb{object. This function simply calls [stats::terms()] on the}formula` object
to extract the predictors of a variable.
}
\examples{
library(lavaan)
data(data_med)
mod <-
"
m ~ a * x + c1 + c2
y ~ b * m + x + c1 + c2
"
fit <- sem(mod, data_med, fixed.x = FALSE)
fit_list <- lm_from_lavaan_list(fit)
terms(fit_list$m)
terms(fit_list$y)

}
\seealso{
\link{terms.object}, \code{\link[=lm_from_lavaan_list]{lm_from_lavaan_list()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
