% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_2_and_3.R
\docType{data}
\name{two_moderators}
\alias{two_moderators}
\title{Two Moderators: Sample Dataset}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{x}{Predictor. Numeric.}
\item{w1}{Moderator 1. Numeric.}
\item{w2}{Moderator 2. Numeric.}
\item{y}{Outcome variable. Numeric.}
\item{c1}{Control variable. variable. Numeric.}
\item{c2}{Control variable. Numeric.}
}
}
\usage{
two_moderators
}
\description{
A two-moderator model.
}
\examples{
library(lavaan)
data(two_moderators)
two_moderators$xw1 <- two_moderators$x * two_moderators$w1
two_moderators$xw2 <- two_moderators$x * two_moderators$w2
mod <-
"
y ~ a * x + w1 + w2 + d1 * xw1 + d2 * xw2 + c1 + c2
w1 ~~ v_w1 * w1
w1 ~ m_w1 * 1
w2 ~~ v_w2 * w2
w2 ~ m_w2 * 1
a_lolo := a + d1 * (m_w1 - sqrt(v_w1)) + d2 * (m_w2 - sqrt(v_w2))
a_lohi := a + d1 * (m_w1 - sqrt(v_w1)) + d2 * (m_w2 + sqrt(v_w2))
a_hilo := a + d1 * (m_w1 + sqrt(v_w1)) + d2 * (m_w2 - sqrt(v_w2))
a_hihi := a + d1 * (m_w1 + sqrt(v_w1)) + d2 * (m_w2 + sqrt(v_w2))
"
fit <- sem(mod, two_moderators, fixed.x = FALSE)
parameterEstimates(fit)[c(1, 4, 5, 8:11, 34:37), ]
}
\keyword{datasets}
