% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path.R
\name{check_path}
\alias{check_path}
\title{Check Whether an Indirect Path is Valid in a Model}
\usage{
check_path(x, y, m = NULL, fit = NULL, est = NULL)
}
\arguments{
\item{x}{Character. The name of predictor at the start of the pathway.}

\item{y}{Character. The name of the outcome variable at
the end of the pathway.}

\item{m}{A vector of the variable names of the
mediators. The pathway goes from the first
mediator successively to the last mediator. If
\code{NULL}, the default, the pathway goes from \code{x}
to \code{y}.}

\item{fit}{The fit object. Currently only supports a
\link[lavaan:lavaan-class]{lavaan::lavaan} object.}

\item{est}{The output of \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. If \code{NULL}, the
default, it will be generated from \code{fit}. If supplied,
\code{fit} will ge ignored.}
}
\value{
Boolean. \code{TRUE} is the path is valid, \code{FALSE} if the
path is invalid.
}
\description{
It checks whether an indirect path exists
in a model.
}
\details{
It checks whether the path defined by a predictor,
an outcome, and sequence of mediators exists in
a model.

For example, in the following model in \link{lavaan} syntax

\if{html}{\out{<div class="sourceCode">}}\preformatted{m1 ~ x
m2 ~ m1
m3 ~ x
y ~ m2 + m3
}\if{html}{\out{</div>}}

This path is valid: `x = "x", y = "y", m = c("m1", "m2")``

This path is invalid: `x = "x", y = "y", m = c("m2")``

This path is also invalid: `x = "x", y = "y", m = c("m1", "m2")``
}
\examples{

library(lavaan)
data(data_serial_parallel)
dat <- data_serial_parallel
mod <-
"
m11 ~ x + c1 + c2
m12 ~ m11 + x + c1 + c2
m2 ~ x + c1 + c2
y ~ m12 + m2 + m11 + x + c1 + c2
"
fit <- sem(mod, dat,
           meanstructure = TRUE, fixed.x = FALSE)

# The following paths are valid
check_path(x = "x", y = "y", m = c("m11", "m12"), fit = fit)
check_path(x = "x", y = "y", m = "m2", fit = fit)
# The following paths are invalid
check_path(x = "x", y = "y", m = c("m11", "m2"), fit = fit)
check_path(x = "x", y = "y", m = c("m12", "m11"), fit = fit)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
