% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path.R
\name{check_path}
\alias{check_path}
\title{One Line Title}
\usage{
check_path(x, y, m = NULL, fit = NULL, est = NULL)
}
\arguments{
\item{x}{Character. The name of predictor at the start of the pathway.}

\item{y}{Character. The name of the outcome variable at
the end of the pathway.}

\item{m}{A vector of the variable names of the
mediators. The pathway goes from the first
mediator successively to the last mediator. If
\code{NULL}, the default, the pathway goes from \code{x}
to \code{y}.}

\item{fit}{The fit object. Currently only supports a
\link[lavaan:lavaan-class]{lavaan::lavaan} object.}

\item{est}{The output of \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. If \code{NULL}, the
default, it will be generated from \code{fit}. If supplied,
\code{fit} will ge ignored.}
}
\value{
Specify what are returned.
}
\description{
One paragraph description
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{
\donttest{
}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
