% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_indirect.R
\name{coef.indirect}
\alias{coef.indirect}
\title{Indirect effect or conditional indirect effect}
\usage{
\method{coef}{indirect}(object, ...)
}
\arguments{
\item{object}{The output of \code{\link[=indirect_i]{indirect_i()}} or
\code{\link[=cond_indirect]{cond_indirect()}}.}

\item{...}{Optional arguments. Ignored by the function.}
}
\value{
A scalar: The estimate of the standardized moderation effect.
}
\description{
Return the estimate of the indirect effect
in the output of \code{\link[=indirect_i]{indirect_i()}} or
\code{\link[=cond_indirect]{cond_indirect()}}.
}
\details{
It just extracts and returns the element \code{indirect}.
}
\examples{


library(lavaan)
dat <- modmed_x1m3w4y1
mod <-
"
m1 ~ a1 * x   + b1 * w1 + d1 * x:w1
m2 ~ a2 * m1  + b2 * w2 + d2 * m1:w2
m3 ~ a3 * m2  + b3 * w3 + d3 * m2:w3
y  ~ a4 * m3  + b4 * w4 + d4 * m3:w4
"
fit <- sem(mod, dat, meanstructure = TRUE, fixed.x = FALSE, se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

wvalues <- c(w1 = 5, w2 = 4, w3 = 2, w4 = 3)

indirect_1 <- indirect_i(x = "x", y = "y", m = c("m1", "m2", "m3"), fit = fit,
                       wvalues = wvalues)
indirect_2 <- (est[est$label == "a1", "est"] +
                wvalues["w1"] * est[est$label == "d1", "est"]) *
              (est[est$label == "a2", "est"] +
                wvalues["w2"] * est[est$label == "d2", "est"]) *
              (est[est$label == "a3", "est"] +
                wvalues["w3"] * est[est$label == "d3", "est"]) *
              (est[est$label == "a4", "est"] +
                wvalues["w4"] * est[est$label == "d4", "est"])
indirect_1$indirect
indirect_2

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
