% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_indirect.R
\name{cond_indirect}
\alias{cond_indirect}
\title{One Line Title}
\usage{
cond_indirect(
  x,
  y,
  m = NULL,
  fit = NULL,
  est = NULL,
  implied_stats = NULL,
  wvalues = NULL,
  standardized_x = FALSE,
  standardized_y = FALSE,
  boot_ci = FALSE,
  level = 0.95,
  boot_out = NULL,
  R = 100,
  seed = NULL
)
}
\arguments{
\item{x}{Character. The name of predictor at the start of the pathway.}

\item{y}{Character. The name of the outcome variable at
the end of the pathway.}

\item{m}{A vector of the variable names of the
moderators. The pathway goes from the first
mediator successively to the last mediator. If
\code{NULL}, the default, the pathway goes from \code{x}
to \code{y}.}

\item{fit}{The fit object. Can be a
\link[lavaan:lavaan-class]{lavaan::lavaan} object or a list of \code{\link[=lm]{lm()}} outputs.}

\item{est}{The output of \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. If \code{NULL}, the
default, it will be generated from \code{fit}. If supplied,
\code{fit} will ge ignored.}

\item{implied_stats}{Implied means, variances, and
covariances of observed variables, of the
form of the output of
\code{\link[lavaan:lavInspect]{lavaan::lavInspect()}} with \code{what} set to
\code{"implied"}. The standard deviations are
extracted from this object for
standardization. Default is \code{NULL}, and
implied statistics will be computed from \code{fit} if required.}

\item{wvalues}{A numeric vector of named elements. The names are the variable
names of the moderators, and the values are the values to
which the moderators will be set to. Default is \code{NULL}.}

\item{standardized_x}{Logical. Whether \code{x} will be standardized. Default is
\code{FALSE}.}

\item{standardized_y}{Logical. Whether \code{y} will be standardized. Default is
\code{FALSE}.}

\item{boot_ci}{Logical. Whether bootstrap confidence interval will be formed.
Default is \code{FALSE}.}

\item{level}{The level of confidence for the bootstrap confidence interval.
Default is .95.}

\item{boot_out}{If \code{boot_ci} is \code{TRUE}, users can supply pregenerated
bootstrap results. This can be the output of \code{\link[=fit2boot_out]{fit2boot_out()}}
or \code{\link[=lm2boot_out]{lm2boot_out()}}. If not supplied, the function will try
to generate them from \code{fit}.}

\item{R}{Integer. If \code{boot_ci} is \code{TRUE}, \code{fit} is a list of \code{\link[=lm]{lm()}} outputs,
and \code{boot_out} is \code{NULL}, this function will do bootstrapping on
\code{fit}. \code{R} is the number of bootstrap samples. Default is 100.}

\item{seed}{If \code{boot_ci} is \code{TRUE}, \code{fit} is a list of \code{\link[=lm]{lm()}} outputs,
and \code{boot_out} is \code{NULL}, this function will do bootstrapping on
\code{fit}. This is the seed for the bootstrapping.
Default is \code{NULL} and seed is not set.}

\item{wlevels}{The output of \code{\link[=merge_mod_levesl]{merge_mod_levesl()}}.}

\item{...}{Arguments to be passed to \code{\link[=cond_indirect]{cond_indirect()}}}
}
\value{
Specify what are returned.
}
\description{
One paragraph description
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{

library(lavaan)
dat <- modmed_x1m3w4y1
mod <-
"
m1 ~ a1 * x   + b1 * w1 + d1 * x:w1
m2 ~ a2 * m1  + b2 * w2 + d2 * m1:w2
m3 ~ a3 * m2  + b3 * w3 + d3 * m2:w3
y  ~ a4 * m3  + b4 * w4 + d4 * m3:w4
"
fit <- sem(mod, dat, meanstructure = TRUE, fixed.x = FALSE, se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

wvalues <- c(w1 = 5, w2 = 4, w3 = 2, w4 = 3)

indirect_1 <- indirect(x = "x", y = "y", m = c("m1", "m2", "m3"), fit = fit,
                       wvalues = wvalues)
indirect_2 <- (est[est$label == "a1", "est"] +
                wvalues["w1"] * est[est$label == "d1", "est"]) *
              (est[est$label == "a2", "est"] +
                wvalues["w2"] * est[est$label == "d2", "est"]) *
              (est[est$label == "a3", "est"] +
                wvalues["w3"] * est[est$label == "d3", "est"]) *
              (est[est$label == "a4", "est"] +
                wvalues["w4"] * est[est$label == "d4", "est"])
indirect_1$indirect
indirect_2

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
