% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_indirect_diff.R
\name{confint.cond_indirect_diff}
\alias{confint.cond_indirect_diff}
\title{Confidence Interval of the
Output of 'cond_indirect_diff()'}
\usage{
\method{confint}{cond_indirect_diff}(object, parm, level = NULL, ...)
}
\arguments{
\item{object}{The output of
\code{\link[=cond_indirect_diff]{cond_indirect_diff()}}.}

\item{parm}{Ignored.}

\item{level}{If set to \code{NULL}, the default,
then the
level of confidence used to generate
\code{object}
is used. If set to a value,
this value will be used
to recompute the confidence intervals.
If the confidence interval is to be
computed from the standard error,
and so \code{level} is not set in \code{object},
then the default value is .95.
(This new behavior applies to 0.3.6.15
and later version.)}

\item{...}{Optional arguments.
Ignored.}
}
\value{
A one-row-two-column data
frame of the confidence limits. If
confidence interval is not available,
the limits are \code{NA}s.
}
\description{
Extract the confidence
interval the output of
\code{\link[=cond_indirect_diff]{cond_indirect_diff()}}.
}
\details{
The \code{confint} method of the
\code{cond_indirect_diff}-class object.

The type of confidence intervals
depends on the call used to
create the object. This function
merely extracts the stored
confidence intervals.
}
