% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_cond_indirect_effects.R
\name{confint.cond_indirect_effects}
\alias{confint.cond_indirect_effects}
\title{Confidence Intervals of Indirect Effects or Conditional
Indirect Effects}
\usage{
\method{confint}{cond_indirect_effects}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{The output of \code{\link[=cond_indirect_effects]{cond_indirect_effects()}}.}

\item{parm}{Ignored. Always returns the bootstrap confidence interval
of the effects for all levels stored.}

\item{level}{The level of confidence, default is .95, returning the
95\% confidence interval. Ignored for now and will
use the level of the stored intervals.}

\item{...}{Additional arguments. Ignored by the function.}
}
\description{
Return the confidence intervals of the conditional
indirect effects or conditional effects in the output of
\code{\link[=cond_indirect_effects]{cond_indirect_effects()}}.
}
\details{
It extracts and returns the columns for confidence
intervals, if available.
}
\examples{

library(lavaan)
dat <- modmed_x1m3w4y1
mod <-
"
m1 ~ x  + w1 + x:w1
m2 ~ m1
y  ~ m2 + x + w4 + m2:w4
"
fit <- sem(mod, dat, meanstructure = TRUE, fixed.x = FALSE, se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

# Examples for cond_indirect():

# Create levels of w1 and w4
w1levels <- mod_levels("w1", fit = fit)
w1levels
w4levels <- mod_levels("w4", fit = fit)
w4levels
w1w4levels <- merge_mod_levels(w1levels, w4levels)

# Conditional effects from x to m1 when w1 is equal to each of the levels
# R should be at least 2000 or 5000 in real research.
out1 <- suppressWarnings(cond_indirect_effects(x = "x", y = "m1",
                      wlevels = w1levels, fit = fit,
                      boot_ci = TRUE, R = 20, seed = 54151,
                      parallel = FALSE))
confint(out1)

# Conditional Indirect effect from x1 through m1 to y,
# when w1 is equal to each of the levels
# R should be at least 2000 or 5000 in real research.
out2 <- suppressWarnings(cond_indirect_effects(x = "x", y = "y", m = c("m1", "m2"),
                      wlevels = w1w4levels, fit = fit,
                      boot_ci = TRUE, R = 20, seed = 54151,
                      parallel = FALSE))
confint(out2)

}
\seealso{
\code{\link[=cond_indirect_effects]{cond_indirect_effects()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
