% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_indirect_list.R
\name{confint.indirect_list}
\alias{confint.indirect_list}
\title{Confidence Intervals of
Indirect Effects in an 'indirect_list'
Object}
\usage{
\method{confint}{indirect_list}(object, parm = NULL, level = NULL, ...)
}
\arguments{
\item{object}{The output of
\code{\link[=many_indirect_effects]{many_indirect_effects()}}.}

\item{parm}{Ignored for now.}

\item{level}{If set to \code{NULL}, the default,
then the
level of confidence used to generate
\code{object}
is used. If set to a value,
this value will be used
to recompute the confidence intervals.
If the confidence interval is to be
computed from the standard error,
and so \code{level} is not set in \code{object},
then the default value is .95.
(This new behavior applies to 0.3.6.15
and later version.)}

\item{...}{Additional arguments.
To be passed to \code{\link[=confint.indirect]{confint.indirect()}}.
(This new behavior applies to 0.3.6.15
and later version.)}
}
\value{
A two-column data frame.
The columns are the limits of
the confidence intervals.
}
\description{
Return the
confidence intervals of the indirect
effects
stored in the output of
\code{\link[=many_indirect_effects]{many_indirect_effects()}}.
}
\details{
It extracts and returns the
stored confidence interval
if available.

The type of confidence intervals
depends on the call used to
compute the effects. This function
merely retrieves the stored estimates,
which could be generated by
nonparametric bootstrapping,
Monte Carlo simulation, or other
methods to be supported in
the future, and uses them to form the
percentile confidence interval.
}
\examples{

library(lavaan)
data(data_serial_parallel)
mod <-
"
m11 ~ x + c1 + c2
m12 ~ m11 + x + c1 + c2
m2 ~ x + c1 + c2
y ~ m12 + m2 + m11 + x + c1 + c2
"
fit <- sem(mod, data_serial_parallel,
           fixed.x = FALSE)
# All indirect paths from x to y
paths <- all_indirect_paths(fit,
                           x = "x",
                           y = "y")
paths
# Indirect effect estimates
# R should be 2000 or even 5000 in real research
# parallel should be used in real research.
fit_boot <- do_boot(fit, R = 45, seed = 8974,
                    parallel = FALSE,
                    progress = FALSE)
out <- many_indirect_effects(paths,
                             fit = fit,
                             boot_ci = TRUE,
                             boot_out = fit_boot)
out
confint(out)



}
\seealso{
\code{\link[=many_indirect_effects]{many_indirect_effects()}}
}
