% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_indirect_list.R
\name{confint.indirect_list}
\alias{confint.indirect_list}
\title{Confidence Intervals of
Indirect Effects in an 'indirect_list'
Object}
\usage{
\method{confint}{indirect_list}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{The output of
\code{\link[=many_indirect_effects]{many_indirect_effects()}}.}

\item{parm}{Ignored for now.}

\item{level}{The level of confidence,
default is .95, returning the 95\%
confidence interval.}

\item{...}{Additional arguments.
Ignored by the function.}
}
\value{
A two-column data frame.
The columns are the limits of
the confidence intervals.
}
\description{
Return the  bootstrap
confidence intervals of the indirect
effects
stored in the output of
\code{\link[=many_indirect_effects]{many_indirect_effects()}}.
}
\details{
It extracts and returns the
stored bootstrap confidence interval
if available.
}
\examples{

dat <- modmed_x1m3w4y1
# TODO


}
\seealso{
\code{\link[=many_indirect_effects]{many_indirect_effects()}}
}
