% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_7_mod.R
\docType{data}
\name{data_mod_2x1w}
\alias{data_mod_2x1w}
\title{Sample Dataset: One Moderator on Two Predictors}
\format{
A data frame with 200 rows
and 6 variables:
\describe{
\item{y}{Outcome variable. Numeric.}
\item{x1}{Predictor. Numeric.}
\item{x2}{Predictor. Numeric.}
\item{w}{Moderator. Numeric.}
\item{c1}{Control variable. Numeric.}
\item{c2}{Control variable. Numeric.}
}
}
\usage{
data_mod_2x1w
}
\description{
A one-moderator model,
with moderator affecting the effects
of two predictors.
}
\examples{
library(lavaan)
data(data_mod_2x1w)
lm_out <- lm(y ~ x1*w + x2*w + c1 + c2, data_mod_2x1w)
out1 <- cond_effects(
  wlevels = "w",
  x = "x1",
  fit = lm_out
)
out1
plot(out1, graph_type = "tumble")

out2 <- cond_effects(
  wlevels = "w",
  x = "x2",
  fit = lm_out
)
out2
plot(out2, graph_type = "tumble")
}
\keyword{datasets}
