% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_7_mod_cat_num.R
\docType{data}
\name{data_mod_cat_num_2w}
\alias{data_mod_cat_num_2w}
\title{Sample Dataset: Mixed Moderators}
\format{
A data frame with 600 rows
and 6 variables:
\describe{
\item{y}{Outcome variable. Numeric.}
\item{x}{Predictor. Numeric.}
\item{w}{Moderator 2. Numeric.}
\item{city}{Moderator 1. String: "City A" and "City B".}
\item{c1}{Control variable. Numeric.}
\item{c2}{Control variable. Numeric.}
}
}
\usage{
data_mod_cat_num_2w
}
\description{
A two-moderator model,
one categorical and one numerical.
}
\examples{
data(data_mod_cat_num_2w)
lm_out <- lm(y ~ x*city*w + c1 + c2, data_mod_cat_num_2w)
out <- cond_effects(
  wlevels = c("city", "w"),
  x = "x",
  fit = lm_out
)
out
plot(out, facet_grid_cols = "city", graph_type = "tumble")
}
\keyword{datasets}
