% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2var.R
\name{factor2var}
\alias{factor2var}
\title{One Line Title}
\usage{
factor2var(
  x_value,
  x_contrasts = "contr.treatment",
  prefix = "",
  add_rownames = TRUE
)
}
\arguments{
\item{x_value}{The vector of categorical variable.}

\item{x_contrasts}{The contrast to be used. Default
is \code{constr.treatment}.}

\item{prefix}{The prefix to be added to the variables
to be created. Default is \code{""}.}

\item{add_rownames}{Whether row names will be added
to the output. Default is \code{FALSE}.}
}
\value{
Specify what are returned.
}
\description{
One paragraph description
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{
\donttest{
}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
