% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_mome.R
\name{index_of_mome}
\alias{index_of_mome}
\alias{index_of_momome}
\title{Index of Moderated Mediation and Index of Moderated
Moderated Mediation}
\usage{
index_of_mome(
  x,
  y,
  m = NULL,
  w = NULL,
  fit = NULL,
  boot_ci = FALSE,
  level = 0.95,
  boot_out = NULL,
  R = 100,
  seed = NULL,
  progress = TRUE,
  ...
)

index_of_momome(
  x,
  y,
  m = NULL,
  w = NULL,
  z = NULL,
  fit = NULL,
  boot_ci = FALSE,
  level = 0.95,
  boot_out = NULL,
  R = 100,
  seed = NULL,
  progress = TRUE,
  ...
)
}
\arguments{
\item{x}{Character. The name of the predictor at the start of the
path.}

\item{y}{Character. The name of the outcome variable at the end of
the path.}

\item{m}{A vector of the variable names of the moderators. The
path goes from the first mediator successively to the last
mediator. If \code{NULL}, the default, the path goes from \code{x} to
\code{y}.}

\item{w}{Character. The name of the moderator.}

\item{fit}{The fit object. Can be a
\link[lavaan:lavaan-class]{lavaan::lavaan} object or a list of \code{\link[=lm]{lm()}} outputs.}

\item{boot_ci}{Logical. Whether bootstrap confidence interval will
be formed. Default is \code{FALSE}.}

\item{level}{The level of confidence for the bootstrap confidence
interval. Default is .95.}

\item{boot_out}{If \code{boot_ci} is \code{TRUE}, users can supply
pregenerated bootstrap estimates. This can be the output of
\code{\link[=do_boot]{do_boot()}}. For \code{\link[=cond_indirect_effects]{cond_indirect_effects()}}, this can be
the output of previous call to \code{\link[=cond_indirect_effects]{cond_indirect_effects()}},
\code{\link[=indirect_effect]{indirect_effect()}}, or \code{\link[=cond_indirect]{cond_indirect()}} with
bootstrap estimates. These stored estimates will be reused such that
there is no need to do bootstrapping again. If not supplied, the
function will try to generate them from \code{fit}.}

\item{R}{Integer. If \code{boot_ci} is \code{TRUE}, \code{fit} is a list of \code{\link[=lm]{lm()}}
outputs, and \code{boot_out} is \code{NULL}, this function will do
bootstrapping on \code{fit}. \code{R} is the number of bootstrap samples.
Default is 100.}

\item{seed}{If \code{boot_ci} is \code{TRUE}, \code{fit} is a list of \code{\link[=lm]{lm()}}
outputs, and \code{boot_out} is \code{NULL}, this function will do
bootstrapping on \code{fit}. This is the seed for the bootstrapping.
Default is \code{NULL} and seed is not set.}

\item{progress}{Logical. Display progress or not. Default is \code{TRUE}.}

\item{...}{Arguments to be passed to \code{\link[=cond_indirect_effects]{cond_indirect_effects()}}}

\item{z}{Character. The name of the second moderator, for computing
the index of moderated moderated mediation.}
}
\value{
It returns an \code{index_of_mome}-class object. This class has a
\code{print} method (\code{\link[=print.index_of_mome]{print.index_of_mome()}}), a \code{coef} method
(\code{\link[=coef.index_of_mome]{coef.index_of_mome()}}), and a \code{confint} method
(\code{\link[=confint.index_of_mome]{confint.index_of_mome()}}).
}
\description{
Compute the index of moderated mediation and
the index of moderated moderated mediation.
}
\details{
TODO
}
\section{Functions}{
\itemize{
\item \code{index_of_mome}: Compute the index of
moderated mediation.

\item \code{index_of_momome}: Compute the index of
moderated moderated mediation.
}}

\examples{

# TOUPDATE
library(lavaan)
dat <- modmed_x1m3w4y1
dat$xw1 <- dat$x * dat$w1
mod <-
"
m1 ~ a * x  + f * w1 + d * xw1
y  ~ b * m1 + cp * x
"
fit <- sem(mod, dat,
           meanstructure = TRUE, fixed.x = FALSE,
           se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

# Create levels of w1, the moderators
w1levels <- mod_levels("w1", fit = fit)
w1levels

# Conditional effects from x to y when w1 is equal to each of the levels
boot_out <- fit2boot_out_do_boot(fit, R = 50, seed = 4314)
out <- cond_indirect_effects(x = "x", y = "y", m = "m1",
                      wlevels = w1levels, fit = fit,
                      boot_ci = TRUE, boot_out = boot_out)
out
out_ind <- index_of_mome(out, from = 2, to = 1)
out_ind
coef(out_ind)
confint(out_ind)



}
\references{
Hayes, A. F. (2015). An index and test of linear moderated mediation.
\emph{Multivariate Behavioral Research, 50}(1), 1-22.
\doi{10.1080/00273171.2014.962683}

Hayes, A. F. (2018). Partial, conditional, and moderated moderated
mediation: Quantification, inference, and interpretation.
\emph{Communication Monographs, 85}(1), 4-40.
\doi{10.1080/03637751.2017.1352100}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
