% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_of_mome.R
\name{index_of_mome}
\alias{index_of_mome}
\title{Generalized Index of Moderated-Mediation}
\usage{
index_of_mome(output, from = NULL, to = NULL, level = 0.95)
}
\arguments{
\item{output}{A \code{cond_indirect_effects}-class object: The output
of \code{\link[=cond_indirect_effects]{cond_indirect_effects()}}.}

\item{from}{A row number of \code{output}.}

\item{to}{A row number of \code{output}. The change in indirect effects
is computed by the change in the level(s) of the moderator(s)
from Row \code{from} to Row \code{to}.}

\item{level}{The level of confidence for the bootstrap confidence
interval. Default is .95.}
}
\value{
It returns an \code{index_of_mome}-class object. This class
has a \code{print} method (\code{\link[=print.index_of_mome]{print.index_of_mome()}}), a \code{coef} method
(\code{\link[=coef.index_of_mome]{coef.index_of_mome()}}), and a \code{confint} method (\code{\link[=confint.index_of_mome]{confint.index_of_mome()}}).
}
\description{
Compute a generalized version of the index of
moderated-mediation proposed by Hayes (2015).
}
\details{
Ths function computes a generalized version of the index of
moderated-mediation
proposed by Hayes (2015).

The variant proposed by Hayes is the change in the indirect effect
along a moderated pathway when a moderator increases by one unit.

The generalized version is the change in the indirect effect
along a moderated pathway when a moderator increases by an amount
meaningful in a context.

For example, the function can compute the change in the indirect effect
when a moderator increases by one standard deviation. We call this
variant the \emph{standardized index of moderated-mediation}.

The function can also compute the change in the standardized
indirect effect when a moderator increases by one unit. We call this
variant the \emph{index of standardized moderated-mediation}.

The change in the standardized
indirect effect when a moderator increases by one standardized deviation
is called the \emph{standardized index of standardized moderated-mediation}.

This function itself does not set the levels of comparison. The levels
to be compared need to be set when calling \code{\link[=cond_indirect_effects]{cond_indirect_effects()}}.
This function extracts required information from the output of
\code{\link[=cond_indirect_effects]{cond_indirect_effects()}}.
}
\section{Functions}{
\itemize{
\item \code{index_of_mome}: Compute the generalized index of
moderated-mediation.
}}

\examples{

library(lavaan)
dat <- modmed_x1m3w4y1
dat$xw1 <- dat$x * dat$w1
mod <-
"
m1 ~ a * x  + f * w1 + d * xw1
y  ~ b * m1 + cp * x
"
fit <- sem(mod, dat, meanstructure = TRUE, fixed.x = FALSE, se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

# Create levels of w1, the moderators
w1levels <- mod_levels("w1", fit = fit)
w1levels

# Conditional effects from x to y when w1 is equal to each of the levels
boot_out <- fit2boot_out_do_boot(fit, R = 50, seed = 4314)
out <- cond_indirect_effects(x = "x", y = "y", m = "m1",
                      wlevels = w1levels, fit = fit,
                      boot_ci = TRUE, boot_out = boot_out)
out
out_ind <- index_of_mome(out, from = 2, to = 1)
out_ind
coef(out_ind)
confint(out_ind)



}
\references{
Hayes, A. F. (2015). An index and test of linear moderated mediation.
\emph{Multivariate Behavioral Research, 50}(1), 1-22.
\doi{10.1080/00273171.2014.962683}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
