% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_mome.R
\name{index_of_mome}
\alias{index_of_mome}
\alias{index_of_momome}
\title{Index of Moderated Mediation and Index of Moderated
Moderated Mediation}
\usage{
index_of_mome(
  x,
  y,
  m = NULL,
  w = NULL,
  fit = NULL,
  boot_ci = FALSE,
  level = 0.95,
  boot_out = NULL,
  R = 100,
  seed = NULL,
  progress = TRUE,
  ...
)

index_of_momome(
  x,
  y,
  m = NULL,
  w = NULL,
  z = NULL,
  fit = NULL,
  boot_ci = FALSE,
  level = 0.95,
  boot_out = NULL,
  R = 100,
  seed = NULL,
  progress = TRUE,
  ...
)
}
\arguments{
\item{x}{Character. The name of the predictor at the start of the
path.}

\item{y}{Character. The name of the outcome variable at the end of
the path.}

\item{m}{A vector of the variable names of the moderators. The
path goes from the first mediator successively to the last
mediator. If \code{NULL}, the default, the path goes from \code{x} to
\code{y}.}

\item{w}{Character. The name of the moderator.}

\item{fit}{The fit object. Can be a
\link[lavaan:lavaan-class]{lavaan::lavaan} object or a list of \code{\link[=lm]{lm()}} outputs.}

\item{boot_ci}{Logical. Whether bootstrap confidence interval will
be formed. Default is \code{FALSE}.}

\item{level}{The level of confidence for the bootstrap confidence
interval. Default is .95.}

\item{boot_out}{If \code{boot_ci} is \code{TRUE}, users can supply
pregenerated bootstrap estimates. This can be the output of
\code{\link[=do_boot]{do_boot()}}. For \code{\link[=cond_indirect_effects]{cond_indirect_effects()}}, this can be
the output of previous call to \code{\link[=cond_indirect_effects]{cond_indirect_effects()}},
\code{\link[=indirect_effect]{indirect_effect()}}, or \code{\link[=cond_indirect]{cond_indirect()}} with
bootstrap estimates. These stored estimates will be reused such that
there is no need to do bootstrapping again. If not supplied, the
function will try to generate them from \code{fit}.}

\item{R}{Integer. If \code{boot_ci} is \code{TRUE}, \code{fit} is a list of \code{\link[=lm]{lm()}}
outputs, and \code{boot_out} is \code{NULL}, this function will do
bootstrapping on \code{fit}. \code{R} is the number of bootstrap samples.
Default is 100.}

\item{seed}{If \code{boot_ci} is \code{TRUE}, \code{fit} is a list of \code{\link[=lm]{lm()}}
outputs, and \code{boot_out} is \code{NULL}, this function will do
bootstrapping on \code{fit}. This is the seed for the bootstrapping.
Default is \code{NULL} and seed is not set.}

\item{progress}{Logical. Display progress or not. Default is \code{TRUE}.}

\item{...}{Arguments to be passed to \code{\link[=cond_indirect_effects]{cond_indirect_effects()}}}

\item{z}{Character. The name of the second moderator, for computing
the index of moderated moderated mediation.}
}
\value{
It returns an \code{cond_indirect_diff}-class object. This class has a
\code{print} method (\code{\link[=print.cond_indirect_diff]{print.cond_indirect_diff()}}), a \code{coef} method
(\code{\link[=coef.cond_indirect_diff]{coef.cond_indirect_diff()}}), and a \code{confint} method
(\code{\link[=confint.cond_indirect_diff]{confint.cond_indirect_diff()}}).
}
\description{
Compute the index of moderated mediation and
the index of moderated moderated mediation.
}
\details{
The function \code{\link[=index_of_mome]{index_of_mome()}} computes the index of moderated
mediation proposed by Hayes. It supports any path in a model,
with one component path moderated.

The function \code{\link[=index_of_momome]{index_of_momome()}} computes the index of
moderated mediation proposed by Hayes. It supports any path
in a model, with two component paths each moderated by
one moderator.
}
\section{Functions}{
\itemize{
\item \code{index_of_mome}: Compute the index of
moderated mediation.

\item \code{index_of_momome}: Compute the index of
moderated moderated mediation.
}}

\examples{

\dontrun{
library(lavaan)
dat <- modmed_x1m3w4y1
dat$xw1 <- dat$x * dat$w1
mod <-
"
m1 ~ a * x  + f * w1 + d * xw1
y  ~ b * m1 + cp * x
ind_mome := d * b
"
fit <- sem(mod, dat,
           meanstructure = TRUE, fixed.x = FALSE,
           se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

# R should be at least 2000 or even 5000 in real research.
boot_out <- do_boot(fit, R = 500, seed = 4314, parallel = FALSE)
out_mome <- index_of_mome(x = "x", y = "y", m = "m1", w = "w1",
                          fit = fit,
                          boot_ci = TRUE, boot_out = boot_out)
out_mome
coef(out_mome)
# From lavaan
print(est[19, ], nd = 8)
confint(out_mome)
}



\dontrun{
library(lavaan)
dat <- modmed_x1m3w4y1
dat$xw1 <- dat$x * dat$w1
dat$m1w4 <- dat$m1 * dat$w4
mod <-
"
m1 ~ a * x  + f1 * w1 + d1 * xw1
y  ~ b * m1 + f4 * w4 + d4 * m1w4 + cp * x
ind_momome := d1 * d4
"
fit <- sem(mod, dat,
           meanstructure = TRUE, fixed.x = FALSE,
           se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

# R should be at least 2000 or even 5000 in real research.
boot_out <- do_boot(fit, R = 500, seed = 4314, parallel = FALSE)
out_momome <- index_of_momome(x = "x", y = "y", m = "m1",
                          w = "w1", z = "w4",
                          fit = fit,
                          boot_ci = TRUE, boot_out = boot_out)
out_momome
coef(out_momome)
print(est[32, ], nd = 8)
confint(out_momome)
}

}
\references{
Hayes, A. F. (2015). An index and test of linear moderated mediation.
\emph{Multivariate Behavioral Research, 50}(1), 1-22.
\doi{10.1080/00273171.2014.962683}

Hayes, A. F. (2018). Partial, conditional, and moderated moderated
mediation: Quantification, inference, and interpretation.
\emph{Communication Monographs, 85}(1), 4-40.
\doi{10.1080/03637751.2017.1352100}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
