% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2est_lm.R
\name{lm2boot_out}
\alias{lm2boot_out}
\title{One Line Title}
\usage{
lm2boot_out(outputs, R = 100, seed = NULL)
}
\arguments{
\item{outputs}{A list of \code{lm} class objects.}

\item{R}{The number of bootstrap samples. Default is 100.}

\item{seed}{The seed for the bootstrapping.
Default is \code{NULL} and seed is not set.}
}
\value{
Specify what are returned.
}
\description{
One paragraph description
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{

library(lavaan)
dat <- modmed_x1m3w4y1
mod <-
"
m1 ~ a1 * x
m2 ~ a2 * m1
m3 ~ a3 * m2
y  ~ a4 * m3 + c4 * x
"
fit <- sem(mod, dat, meanstructure = TRUE, fixed.x = FALSE, se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

check_path(x = "x", y = "m3", m = c("m1", "m2"), fit = fit)
check_path(x = "x", y = "y", m = c("m1", "m2"), fit = fit)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
