% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manymome-package.R
\docType{package}
\name{manymome-package}
\alias{manymome}
\alias{manymome-package}
\title{manymome: Moderation, Mediation, and Moderated-Mediation in SEM and Regression}
\description{
Functions for computing indirect effects, cnoditioanl effects, and conditional indirect effects, and their bootstraping confidence intervals. Support standadized (partially or completely) indirect effects, conditional effects, and conditional indirect effects. Model fitting can be done by SEM (using 'sem()' in 'lavaan') or regression (using 'lm()').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/manymome/}
  \item Report bugs at \url{https://github.com/sfcheung/manymome/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Sing-Hang Cheung (\href{https://orcid.org/0000-0001-5182-0752}{ORCID})
}

}
\keyword{internal}
