% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_indirect_effects_math.R
\name{math_indirect}
\alias{math_indirect}
\alias{+.indirect}
\alias{-.indirect}
\title{Math Operators for 'indirect'-Class Objects}
\usage{
\method{+}{indirect}(e1, e2)

\method{-}{indirect}(e1, e2)
}
\arguments{
\item{e1}{A 'indirect'-class object.}

\item{e2}{A 'indirect'-class object.}
}
\value{
A 'indirect'-class object with list of effects stored.
}
\description{
Mathematic operators for
'indirect'-class object.
}
\details{
For now, only has a \code{+} operator and a \code{-} operator.
These operators can be used to estimate and test
a function of effects between the same pair of variables
but along different paths.

For example, they can be used to compute and test the total effects
along different paths. They can also be used to compute and
test the difference between the effects along two paths.

The operators will check whether an operation is valid. A operation
is not valid if (a) the two paths do not start from the same variable,
(b) the two paths do not end at the same variable, (c) a path appears
in both objects, (d) moderators are involved but they are not
set to the same values in both objects, and (e) bootstrap estimates
stored in \code{boot_out} are not identical.
}
\examples{
library(lavaan)
dat <- modmed_x1m3w4y1
mod <-
"
m1 ~ a1 * x  + d1 * w1 + e1 * x:w1
m2 ~ m1 + a2 * x
y  ~ b1 * m1 + b2 * m2 + cp * x
"
fit <- sem(mod, dat, meanstructure = TRUE, fixed.x = FALSE, se = "none", baseline = FALSE)
est <- parameterEstimates(fit)
hi_w1 <- mean(dat$w1) + sd(dat$w1)

# Examples for cond_indirect():

# Conditional effect from x to m1 when w1 is 1 SD above mean
out1 <- cond_indirect(x = "x", y = "y", m = c("m1", "m2"),
              wvalues = c(w1 = hi_w1), fit = fit)
out2 <- cond_indirect(x = "x", y = "y", m = c("m2"),
              wvalues = c(w1 = hi_w1), fit = fit)
out3 <- cond_indirect(x = "x", y = "y",
              wvalues = c(w1 = hi_w1), fit = fit)

out12 <- out1 + out2
out12
out123 <- out1 + out2 + out3
out123
coef(out1) + coef(out2) + coef(out3)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
