% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_levels.R
\name{mod_levels}
\alias{mod_levels}
\alias{mod_levels_list}
\title{One Line Title}
\usage{
mod_levels(
  w,
  fit,
  w_type = c("auto", "numeric", "categorical"),
  w_method = c("sd", "percentile"),
  sd_from_mean = c(-1, 0, 1),
  percentiles = c(0.16, 0.5, 0.84),
  extract_gp_names = TRUE,
  prefix = NULL
)

mod_levels_list(
  ...,
  fit,
  w_type = c("auto", "numeric", "categorical"),
  w_method = c("sd", "percentile"),
  sd_from_mean = c(-1, 0, 1),
  percentiles = c(0.16, 0.5, 0.84),
  extract_gp_names = TRUE,
  prefix = NULL,
  merge = FALSE
)
}
\arguments{
\item{w}{Character. The names of the moderator. If the moderator is
categorical with 3 or more groups, this is the vector of the
indicator variables.}

\item{fit}{The fit object. Can be a
\link[lavaan:lavaan-class]{lavaan::lavaan} object or a list of \code{\link[=lm]{lm()}} outputs.}

\item{w_type}{Character. Whether the moderator is a \code{"numeric"}
variable or a \code{"categorical"} variable. If \code{"auto"},
the function will try to determine the type automatically.}

\item{w_method}{Character, either \code{"sd"} or \code{"percentile"}.
If \code{"sd"}, the levels are defined by the
distance from the mean in terms of standard deviation.
if \code{"percentile"}, the levels are defined in percentiles.}

\item{sd_from_mean}{A numeric vector. Specify the distance in standard
deviation from the mean for each level. Default is
\code{c(-1, 0, 1)}. Ignored if \code{w_method} is not equal to
\code{"sd"}.}

\item{percentiles}{A numeric vector. Specify the percentile (in proportion)
for each level. Default is \code{c(.16, .50, .84)}, corresponding
approximately to one standard deviation below mean,
mean, and one standard deviation above mean in a normal
distribution. Ignored if \code{w_method} is not equal to
\code{"percentile"}.}

\item{extract_gp_names}{Logical. If \code{TRUE}, the default, the function will
try to determine the name of each group from the
variable names.}

\item{prefix}{Character. If \code{extract_gp_names} is \code{TRUE} and \code{prefix} is
supplied, it will be removed from the variable names
to create the group names. Default is \code{NULL}, and the function
will try to determine the prefix automatically.}

\item{...}{The names of moderators variables. For a categorical variable,
it should be a vector of variable names.}

\item{merge}{If \code{TRUE}, \code{\link[=mod_levels_list]{mod_levels_list()}} will call \code{\link[=merge_mod_levels]{merge_mod_levels()}}
and return the merged levels. Default is \code{FALSE}.}
}
\value{
Specify what are returned.
}
\description{
One paragraph description
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\section{Functions}{
\itemize{
\item \code{mod_levels}: Generate levels for one moderaotr.

\item \code{mod_levels_list}: Generate levels for several moderaotrs.
}}

\examples{

library(lavaan)
dat <- modmed_x1m3w4y1
mod <-
"
m1 ~ a1 * x
m2 ~ a2 * m1
m3 ~ a3 * m2
y  ~ a4 * m3 + c4 * x
"
fit <- sem(mod, dat, meanstructure = TRUE, fixed.x = FALSE, se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

check_path(x = "x", y = "m3", m = c("m1", "m2"), fit = fit)
check_path(x = "x", y = "y", m = c("m1", "m2"), fit = fit)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
