% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_mediation_plot.R
\name{plot.q_mediation}
\alias{plot.q_mediation}
\title{Plot Method for the Output
of 'q_mediation' Family}
\usage{
\method{plot}{q_mediation}(x, ...)
}
\arguments{
\item{x}{The output of
\code{\link[=q_mediation]{q_mediation()}},
\code{\link[=q_simple_mediation]{q_simple_mediation()}},
\code{\link[=q_serial_mediation]{q_serial_mediation()}}, and
\code{\link[=q_parallel_mediation]{q_parallel_mediation()}}.
(Named \code{x}
because it is required in the naming
of arguments of the \code{plot} generic
function.)}
}
\value{
A \code{ggplot2} graph. Plotted if
not assigned to a name. It can be
further modified like a usual
\code{ggplot2} graph.
}
\description{
Plot the path model
fitted by the family of 'q_mediation'
functions.
}
\details{
(TODO)
}
