% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_indirect_diff.R
\name{print.cond_indirect_diff}
\alias{print.cond_indirect_diff}
\title{Print the Output of
'cond_indirect_diff'}
\usage{
\method{print}{cond_indirect_diff}(x, digits = 3, pvalue = FALSE, pvalue_digits = 3, ...)
}
\arguments{
\item{x}{The output of
\code{\link[=cond_indirect_diff]{cond_indirect_diff()}}.}

\item{digits}{The number of decimal
places in the printout.}

\item{pvalue}{Logical. If \code{TRUE},
asymmetric \emph{p}-value based on
bootstrapping will be printed if
available. Default is \code{FALSE.}}

\item{pvalue_digits}{Number of decimal
places to display for the \emph{p}-value.
Default is 3.}

\item{...}{Optional arguments.
Ignored.}
}
\value{
It returns \code{x} invisibly.
Called for its side effect.
}
\description{
Print the output of
\code{\link[=cond_indirect_diff]{cond_indirect_diff()}}.
}
\details{
The \code{print} method of the
\code{cond_indirect_diff}-class object.
}
\seealso{
\code{\link[=cond_indirect_diff]{cond_indirect_diff()}}
}
