% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_delta_med.R
\name{print.delta_med}
\alias{print.delta_med}
\title{Print a 'delta_med' Class Object}
\usage{
\method{print}{delta_med}(x, digits = 3, level = NULL, full = FALSE, ...)
}
\arguments{
\item{x}{A \code{delta_med}-class object.}

\item{digits}{The number of digits
after the decimal. Default is 3.}

\item{level}{The level of confidence
of bootstrap confidence interval,
if requested when created. If \code{NULL},
the default, the level requested when
calling \code{\link[=delta_med]{delta_med()}} is used. If
not null, then this level will be
used.}

\item{full}{Logical. Whether
additional information will be printed.
Default is \code{FALSE}.}

\item{...}{Optional arguments.
Ignored.}
}
\value{
\code{x} is returned invisibly. Called
for its side effect.
}
\description{
Print the content of
a \code{delta_med}-class object.
}
\details{
It prints the output of
\code{delta_med()}, which is a
\code{delta_med}-class object.
}
\examples{

# TODO: To prepare
x <- 123
print(x)

}
\seealso{
\code{\link[=delta_med]{delta_med()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
