% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lm_list.R
\name{summary.lm_list}
\alias{summary.lm_list}
\alias{print.summary_lm_list}
\title{Summary of an \code{lm_list}-Class Object}
\usage{
\method{summary}{lm_list}(object, ...)

\method{print}{summary_lm_list}(x, digits = 3, ...)
}
\arguments{
\item{object}{The output of \code{\link[=lm2list]{lm2list()}}.}

\item{...}{Other arguments. Not used.}

\item{x}{An object of class \code{summary_lm_list}.}

\item{digits}{The number of significant digits in printing
numerical results.}
}
\value{
\code{x} is returned invisibly.
}
\description{
The summary of content of the output of \code{\link[=lm2list]{lm2list()}}.
}
\section{Functions}{
\itemize{
\item \code{print.summary_lm_list}: Print method for output of summary for lm_list.
}}

\examples{

data(data_serial_parallel)
lm_m11 <- lm(m11 ~ x + c1 + c2, data_serial_parallel)
lm_m12 <- lm(m12 ~ m11 + x + c1 + c2, data_serial_parallel)
lm_m2 <- lm(m2 ~ x + c1 + c2, data_serial_parallel)
lm_y <- lm(y ~ m11 + m12 + m2 + x + c1 + c2, data_serial_parallel)
# Join them to form a lm_list-class object
lm_serial_parallel <- lm2list(lm_m11, lm_m12, lm_m2, lm_y)
lm_serial_parallel
summary(lm_serial_parallel)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
